requirements.txt
PyYAML==3.11
SQLAlchemy==0.9.7
botocore==0.62.0
bottle==0.12.7
certifi==14.05.14
ipython==2.2.0
jmespath==0.4.1
psycopg2==2.5.4
python-dateutil==2.2
requests==2.4.0
six==1.7.3
